+++ samplebot.js
@@ -14,7 +14,7 @@ import {skills as searchSkills} from './packages/skill-search/index.js';
 //       <div class="flex flex-col space-y-2 text-xs max-w-xs mx-2 order-2 items-start">
 //           <div><span class="px-4 py-2 rounded-lg inline-block rounded-bl-none bg-gray-300 text-gray-600">Can be verified on any platform using docker</span></div>
 //       </div>
-//       <img src="https://images.unsplash.com/photo-1549078642-b2ba4bda0cdb?ixlib=rb-1.2.1&amp;ixid=eyJhcHBfaWQiOjEyMDd9&amp;auto=format&amp;fit=facearea&amp;facepad=3&amp;w=144&amp;h=144" alt="My profile" class="w-6 h-6 rounded-full order-1">
+//       <span class="w-6 h-6 rounded-full order-1">bot</span>
 //     </div>
 // </div>
 // user
@@ -23,7 +23,7 @@ import {skills as searchSkills} from './packages/skill-search/index.js';
 //       <div class="flex flex-col space-y-2 text-xs max-w-xs mx-2 order-1 items-end">
 //           <div><span class="px-4 py-2 rounded-lg inline-block rounded-br-none bg-blue-600 text-white ">Your error message says permission denied, npm global installs must be given root privileges.</span></div>
 //       </div>
-//       <img src="https://images.unsplash.com/photo-1590031905470-a1a1feacbb0b?ixlib=rb-1.2.1&amp;ixid=eyJhcHBfaWQiOjEyMDd9&amp;auto=format&amp;fit=facearea&amp;facepad=3&amp;w=144&amp;h=144" alt="My profile" class="w-6 h-6 rounded-full order-2">
+//       <span class="w-6 h-6 rounded-full order-2">user</span>
 //     </div>
 // </div>
 const renderMessage = (message, charactor, role = 'bot') => {